updating the compliance masonry to work with new schema, and the new masonry.py script
